requirements.txt
setuptools>=39.2.0
PyYAML
